feat: modify manage rental detail page
