+++ packages/web/src/common/components/ProgressStatus.tsx
@@ -11,6 +11,7 @@ import Typography from "@sparcs-clubs/web/common/components/Typography";
 interface ProgressStatusProps {
   labels: string[];
   progress: StatusAndDate[];
+  // TODO: /my까지 수정하면 infoText, optional 지우기
   infoText?: string;
   optional?: React.ReactNode;
 }
@@ -30,14 +31,7 @@ const ProgressStatus: React.FC<ProgressStatusProps> = ({
   optional = undefined,
 }) => (
   <ProgressCheckSectionWrapper>
-    <Typography
-      ff="PRETENDARD"
-      fw="MEDIUM"
-      fs={16}
-      lh={20}
-      color="BLACK"
-      style={{ width: "100%" }}
-    >
+    <Typography fw="MEDIUM" fs={16} lh={20} style={{ width: "100%" }}>
       신청 상태
     </Typography>
     <ProgressCheckSection labels={labels} progress={progress} />
